+++ FitData/Home/Views/HomeView.swift
@@ -7,24 +7,10 @@
 
 import SwiftUI
 
-struct HomeView: View {
-    @State var calories: Int = 123
-    @State var active: Int = 52
-    @State var stand: Int = 8
 
-    var mockActivities = [
-        Activity(id: 0, title: "Today Steps", subtitle: "Goal: 12,000", image: "figure.walk", tintColor: .green, amount: "9812"),
-        Activity(id: 1, title: "Today Steps", subtitle: "Goal: 12,000", image: "figure.walk", tintColor: .yellow, amount: "9012"),
-        Activity(id: 2, title: "Today Steps", subtitle: "Goal: 12,000", image: "figure.walk", tintColor: .red, amount: "4532"),
-        Activity(id: 3, title: "Today Steps", subtitle: "Goal: 12,000", image: "figure.walk", tintColor: .blue, amount: "6894")
-    ]
-    
-    var mockWorkouts = [
-        Workout(id: 0, title: "Running", image: "figure.run", tintColor: .cyan, duration: "38 mins", date: "Aug 10", calories: "783 cals"),
-        Workout(id: 1, title: "Strength Training", image: "figure.run", tintColor: .cyan, duration: "45 mins", date: "Aug 20", calories: "783 cals"),
-        Workout(id: 2, title: "Soccer", image: "figure.run", tintColor: .cyan, duration: "81 mins", date: "Aug 2", calories: "783 cals"),
-        Workout(id: 3, title: "Running", image: "figure.run", tintColor: .cyan, duration: "51 mins", date: "Aug 8", calories: "783 cals")
-    ]
+struct HomeView: View {
+    @StateObject var viewModel = HomeViewModel()
+  
     
     var body: some View {
         NavigationStack {
@@ -38,7 +24,7 @@ struct HomeView: View {
                         Spacer()
                         
                         
-                        VStack {
+                        VStack(alignment: .leading) {
                             VStack(alignment: .leading, spacing: 8) {
                                 Text("Calories")
                                     .font(.callout)
@@ -46,7 +32,7 @@ struct HomeView: View {
                                     .foregroundColor(.red)
                                 
                                 
-                                Text("123 calories")
+                                Text("\(viewModel.calories)")
                                     .bold()
                             }
                             .padding(.bottom)
@@ -58,7 +44,7 @@ struct HomeView: View {
                                     .foregroundColor(.green)
                                 
                                 
-                                Text("52  minutes")
+                                Text("\(viewModel.active)")
                                     .bold()
                             }
                             .padding(.bottom)
@@ -70,7 +56,7 @@ struct HomeView: View {
                                     .foregroundColor(.blue)
                                 
                                 
-                                Text("8 hours")
+                                Text("\(viewModel.stand)")
                                     .bold()
                             }
                         }
@@ -78,10 +64,10 @@ struct HomeView: View {
                         Spacer()
                         
                         ZStack {
-                            ProgressCircleView(progress: $calories, color: .red, goal: 600)
-                            ProgressCircleView(progress: $active, color: .green, goal: 60)
+                            ProgressCircleView(progress: $viewModel.calories, color: .red, goal: 600)
+                            ProgressCircleView(progress: $viewModel.active, color: .green, goal: 60)
                                 .padding(.all, 20)
-                            ProgressCircleView(progress: $stand, color: .blue, goal: 12)
+                            ProgressCircleView(progress: $viewModel.stand, color: .blue, goal: 12)
                                 .padding(.all, 40)
                         }
                         .padding(.horizontal)
@@ -110,7 +96,9 @@ struct HomeView: View {
                     
                     
                     LazyVGrid(columns: Array(repeating: GridItem(spacing: 20), count: 2)) {
-                        ForEach(mockActivities, id: \.id) { activity in
+                        ForEach(viewModel.mockActivities,
+                            id: \.id
+                        ) { activity in
                         ActivityCard(activity: activity)
                         }
                         
@@ -137,7 +125,7 @@ struct HomeView: View {
                     .padding(.top)
                     
                     LazyVStack {
-                        ForEach(mockWorkouts, id: \.id) { workout in
+                        ForEach(viewModel.mockWorkouts, id: \.id) { workout in
                         WorkoutCard(workout: workout)
                         }
                     }
